code/requirements.txt
torch
torchvision
torchaudio
numpy
timm
einops
pyssim
opencv-python
matplotlib
Pillow
